% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-max-weight.R
\name{plot_exercise_max_weight}
\alias{plot_exercise_max_weight}
\title{Plot maximum exercise weight over time}
\usage{
plot_exercise_max_weight(exercise_summary, ncol = 3)
}
\arguments{
\item{exercise_summary}{Exercise-level summary data from
summarize_exercises().}

\item{ncol}{Number of facet columns.}
}
\value{
A ggplot object.
}
\description{
Plot maximum exercise weight over time
}
