% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-volume.R
\name{plot_exercise_volume}
\alias{plot_exercise_volume}
\title{Plot exercise volume over time}
\usage{
plot_exercise_volume(exercise_summary, ncol = 3)
}
\arguments{
\item{exercise_summary}{Exercise-level summary data from
summarize_exercises().}

\item{ncol}{Number of facet columns.}
}
\value{
A ggplot object.
}
\description{
Plot exercise volume over time
}
