% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sets.R
\name{plot_set_performance}
\alias{plot_set_performance}
\title{Plot set-level load and repetitions}
\usage{
plot_set_performance(sets, ncol = 3)
}
\arguments{
\item{sets}{A set-level tibble, usually from read_lifting_sets().}

\item{ncol}{Number of facet columns.}
}
\value{
A ggplot object.
}
\description{
Plot set-level load and repetitions
}
